fix(ui): separating the screens.
