+++ app/src/main/java/com/flowintent/workspace/data/remote/AnalyticsAdapter.kt
