+++ app/src/main/java/com/flowintent/workspace/nav/NavBottomBar.kt
@@ -1,0 +1,39 @@
+package com.flowintent.workspace.nav
+
+import androidx.compose.material3.Icon
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.navigation.NavHostController
+
+@Composable
+fun BottomNavigationBar(startDestination: Navigation, navController: NavHostController) {
+    var selectedDestination by rememberSaveable { mutableIntStateOf(startDestination.ordinal) }
+
+    NavigationBar {
+        Navigation.entries.forEachIndexed { index, navigation ->
+            NavigationBarItem(
+                selected = selectedDestination == index,
+                onClick = {
+                    navController.navigate(route = navigation.route) {
+                        popUpTo(navController.graph.startDestinationId)
+                        launchSingleTop = true
+                    }
+                    selectedDestination = index
+                },
+                icon = {
+                    Icon(
+                        imageVector = navigation.icon,
+                        contentDescription = navigation.contentDescription
+                    )
+                },
+                label = { Text(text = navigation.label) }
+            )
+        }
+    }
+}
