+++ app/src/main/java/com/flowintent/workspace/ui/Reminder.kt
@@ -1,0 +1,23 @@
+package com.flowintent.workspace.ui
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun ReminderScreen() {
+    Column(
+        modifier = Modifier.fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        Text(
+            text = "Reminder",
+            modifier = Modifier.align(alignment = Alignment.CenterHorizontally)
+        )
+    }
+}
